Finished all features for session user. Added a paginate for search results. Added new routes

+++ app/controllers/registrations_controller.rb
@@ -29,17 +29,19 @@ class RegistrationsController < Devise::RegistrationsController
   def build_resource(hash=nil)
     hash ||= resource_params || {}
     self.resource = resource_class.new_with_session(hash, session)
-    if session.has_key?(:session_credential_id)
-      old_user = SessionCredential.find_by_session_id session[:session_credential_id]
-      if old_user.nil?
-        self.resource.author_id = SessionCredential.generate_id
+    unless params[:user].nil? || params[:user][:opt_in].nil? || params[:user][:opt_in] == false
+      if session.has_key?(:session_credential_id)
+        old_user = SessionCredential.find_by_session_id session[:session_credential_id]
+        if old_user.nil?
+          self.resource.author_id = SessionCredential.generate_id
+        else
+          self.resource.author_id = old_user.author_id
+          session[:session_credential_id] = nil
+          old_user.destroy
+        end
       else
-        self.resource.author_id = old_user.author_id
-        session[:session_credential_id] = nil
-        old_user.destroy
+        self.resource.author_id = SessionCredential.generate_id
       end
-    else
-      self.resource.author_id = SessionCredential.generate_id
     end
   end
